+++ src/main/java/com/readum/presentation/controller/userBook/dto/UserBookCreateRequest.java
@@ -1,4 +1,4 @@
-package com.readum.presentation.controller.user.dto;
+package com.readum.presentation.controller.userBook.dto;
 
 import com.readum.domain.userBook.dto.UserBookCreateCommand;
 import io.swagger.v3.oas.annotations.media.Schema;
